+++ tests/Unit/Benchmark/Executor/ParameterSetExtractorTest.php
@@ -29,6 +29,16 @@ class ParameterSetExtractorTest extends TestCase
         $this->provideParams(['provideParams']);
     }
 
+    public function testProvideParameterFromIterator()
+    {
+        $this->provideParams(['provideIterator']);
+    }
+
+    public function testProvideParameterFromGenerator()
+    {
+        $this->provideParams(['provideGenerator']);
+    }
+
     public function testThrowsExceptionIfParameterDoesntExist()
     {
         $this->expectException(ScriptErrorException::class);